Improve module-map
It now takes a payload type as template parameter. This allows to have different payloads for modules read from ETL and perf.data files.
