+++ snail/analysis/detail/etl_file_process_context.hpp
@@ -70,6 +70,16 @@ public:
         }
     };
 
+    struct module_data
+    {
+        std::string filename;
+
+        [[nodiscard]] friend bool operator==(const module_data& lhs, const module_data& rhs)
+        {
+            return lhs.filename == rhs.filename;
+        }
+    };
+
     using process_history = detail::history<process_id_t, timestamp_t, process_data>;
     using thread_history  = detail::history<thread_id_t, timestamp_t, thread_data>;
 
@@ -92,7 +102,7 @@ public:
 
     const std::set<std::pair<thread_id_t, timestamp_t>>& get_process_threads(process_id_t process_id) const;
 
-    const module_map& get_modules(process_id_t process_id) const;
+    const module_map<module_data>& get_modules(process_id_t process_id) const;
 
     const std::vector<sample_info>& process_samples(process_id_t process_id) const;
 
@@ -129,7 +139,7 @@ private:
 
     std::unordered_map<process_id_t, profiler_process_info> profiler_processes_;
 
-    std::unordered_map<process_id_t, module_map> modules_per_process;
+    std::unordered_map<process_id_t, module_map<module_data>> modules_per_process;
 
     std::unordered_map<process_id_t, std::vector<sample_info>> samples_per_process;
