+++ snail/analysis/perf_data_data_provider.cpp
@@ -43,16 +43,16 @@ struct perf_data_sample_data : public sample_data
             const auto& symbol = (module == nullptr) ?
                                      resolver->make_generic_symbol(instruction_pointer) :
                                      resolver->resolve_symbol(detail::dwarf_resolver::module_info{
-                                                                  .image_filename = module->file_name,
+                                                                  .image_filename = module->payload.filename,
                                                                   .image_base     = module->base,
-                                                                  .page_offset    = module->page_offset,
+                                                                  .page_offset    = module->payload.page_offset,
                                                                   .process_id     = process_id,
                                                                   .load_timestamp = load_timestamp},
                                                               instruction_pointer);
 
             co_yield stack_frame{
                 .symbol_name             = symbol.name,
-                .module_name             = module == nullptr ? unkown_module_name : module->file_name,
+                .module_name             = module == nullptr ? unkown_module_name : module->payload.filename,
                 .file_path               = symbol.file_path,
                 .function_line_number    = symbol.function_line_number,
                 .instruction_line_number = symbol.instruction_line_number};
